[housekeeping] fix merge conflicts
