--- src/routes/user.route.js
@@ -1,97 +1,0 @@
-import { Router } from 'express';
-
-import { createUser, confirmUser } from './controllers/users';
-import { resetPassword, forgotPassword } from './controllers/password';
-import { validUser } from './middlewares/validateInput';
-const userRouter = Router();
-
-/**
- * @swagger
- *
- * /api/v1/users:
- *   post:
- *     description: User Registration Endpoint
- *     produces:
- *       - application/json
- *     request:
- *         content:
- *         - application/json
- *         schema:
- *           type: array
- *           items:
- *         $ref: '#/definitions/users'
- *     responses:
- *       201:
- *         description: User created
- *       400:
- *         description: Bad request.
- *       401:
- *         description: Authorization information is missing or invalid.
- *       404:
- *        description: A user with the specified ID was not found.
- *       5XX:
- *        description: Unexpected error.
- */
-userRouter.post('/', validUser, createUser);
-userRouter.get('/confirmation/:token', confirmUser);
-/**
- * @swagger
- *
- * /api/v1/users:
- *   post:
- *     description: Reset Password
- *     produces:
- *       - application/json
- *     request:
- *         content:
- *         - application/json
- *         schema:
- *           type: array
- *           items:
- *         $ref: '#/definitions/users'
- *     responses:
- *       201:
- *         description: User created
- *       400:
- *         description: Bad request.
- *       401:
- *         description: Authorization information is missing or invalid.
- *       404:
- *        description: A user with the specified ID was not found.
- *       5XX:
- *        description: Unexpected error.
- */
-
-userRouter.patch('/reset-password/:token', resetPassword);
-
-/**
- * @swagger
- *
- * /api/v1/users:
- *   post:
- *     description: Forgot Password
- *     produces:
- *       - application/json
- *     request:
- *         content:
- *         - application/json
- *         schema:
- *           type: array
- *           items:
- *         $ref: '#/definitions/users'
- *     responses:
- *       201:
- *         description: User created
- *       400:
- *         description: Bad request.
- *       401:
- *         description: Authorization information is missing or invalid.
- *       404:
- *        description: A user with the specified ID was not found.
- *       5XX:
- *        description: Unexpected error.
- */
-
-userRouter.post('/forgot-password', forgotPassword);
-
-export default userRouter;
